// src/components/default/blocks/TaskTemplate/types.ts
import { GridPaginationModel } from '@mui/x-data-grid';
import { ChangeEvent, KeyboardEvent, MouseEvent } from 'react';

interface IDateFilterValues {
  start?: string | null;
  end?: string | null;
}

interface ITaskTemplatesFilter {
  data: number[] | string | IDateFilterValues;
  mode?: FilterMode;
}

interface ITaskTemplatesDateFilter {
  data: IDateFilterValues;
  mode?: FilterMode;
}

export interface IGuidesInfo {
  data: {
    name: string;
    id: number | string;
    checklist_items: IChecklistItemType[];
    tool?: ITool | {};
  };
}

export interface ITaskTemplatesFilters {
  step_templates?: ITaskTemplatesFilter;
  professions?: ITaskTemplatesFilter;
  frequencies?: ITaskTemplatesFilter;
  actions?: ITaskTemplatesFilter;
  objects?: ITaskTemplatesFilter;
  tools?: ITaskTemplatesFilter;
  created_by?: ITaskTemplatesFilter;
  createdAt?: ITaskTemplatesDateFilter;
  is_draft?: boolean | string;
}

export interface IViewProps {
  rows: Array<ITaskTemplate>;
  view: string;
  searchValue: string;
  paginationModel: GridPaginationModel;
  commonData?: ICommonData;
  isSmall?: boolean;
  cardsStorageActive: boolean;
  handleCardsStorage: (value: boolean) => void;
  toggleFilters: (open: boolean) => (event: KeyboardEvent | MouseEvent) => void;
  handleChangeView: (_: MouseEvent<HTMLElement>, newView: string) => void;
  handleActions: (visible: boolean, id?: number | null) => void;
  handleSearch: (e: ChangeEvent<HTMLInputElement>) => void;
  handlePagination: (newPaginationModel: GridPaginationModel) => void;
  handleSetModal: (data: any) => void;
}